keep bad reads in output
Previously, mia tried to filter badly aligning stuff from the set of
reads.  Now it just keeps bad alignments from contributing to the
consensus, but does pass them to the output.  It is hoped that this
gives a good assembly while allowed for a less badly biased
contamination check afterwards.